+++ pkg/conduit/runtime.go
@@ -45,6 +45,8 @@ import (
 	"github.com/conduitio/conduit/pkg/foundation/metrics"
 	"github.com/conduitio/conduit/pkg/foundation/metrics/measure"
 	"github.com/conduitio/conduit/pkg/foundation/metrics/prometheus"
+	"github.com/conduitio/conduit/pkg/http/api"
+	"github.com/conduitio/conduit/pkg/http/openapi"
 	"github.com/conduitio/conduit/pkg/lifecycle"
 	lifecycle_v2 "github.com/conduitio/conduit/pkg/lifecycle-poc"
 	"github.com/conduitio/conduit/pkg/orchestrator"
@@ -60,8 +62,6 @@ import (
 	"github.com/conduitio/conduit/pkg/processor"
 	"github.com/conduitio/conduit/pkg/provisioning"
 	"github.com/conduitio/conduit/pkg/schemaregistry"
-	"github.com/conduitio/conduit/pkg/web/api"
-	"github.com/conduitio/conduit/pkg/web/openapi"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	grpcruntime "github.com/grpc-ecosystem/grpc-gateway/v2/runtime"
 	"github.com/piotrkowalczuk/promgrpc/v4"
