+++ pkg/http/api/connector_v1_test.go
@@ -28,9 +28,9 @@ import (
 	"github.com/conduitio/conduit/pkg/connector"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/foundation/log"
+	apimock "github.com/conduitio/conduit/pkg/http/api/mock"
+	"github.com/conduitio/conduit/pkg/http/api/toproto"
 	"github.com/conduitio/conduit/pkg/inspector"
-	apimock "github.com/conduitio/conduit/pkg/web/api/mock"
-	"github.com/conduitio/conduit/pkg/web/api/toproto"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	"github.com/google/uuid"
 	"github.com/matryer/is"
