+++ pkg/http/api/pipeline_v1.go
@@ -21,10 +21,10 @@ import (
 	"regexp"
 
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
+	"github.com/conduitio/conduit/pkg/http/api/fromproto"
+	"github.com/conduitio/conduit/pkg/http/api/status"
+	"github.com/conduitio/conduit/pkg/http/api/toproto"
 	"github.com/conduitio/conduit/pkg/pipeline"
-	"github.com/conduitio/conduit/pkg/web/api/fromproto"
-	"github.com/conduitio/conduit/pkg/web/api/status"
-	"github.com/conduitio/conduit/pkg/web/api/toproto"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	"google.golang.org/grpc"
 )
