+++ pkg/http/api/processor_v1.go
@@ -27,11 +27,11 @@ import (
 	opencdcv1 "github.com/conduitio/conduit-commons/proto/opencdc/v1"
 	processorSdk "github.com/conduitio/conduit-processor-sdk"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
+	"github.com/conduitio/conduit/pkg/http/api/fromproto"
+	"github.com/conduitio/conduit/pkg/http/api/status"
+	"github.com/conduitio/conduit/pkg/http/api/toproto"
 	"github.com/conduitio/conduit/pkg/inspector"
 	"github.com/conduitio/conduit/pkg/processor"
-	"github.com/conduitio/conduit/pkg/web/api/fromproto"
-	"github.com/conduitio/conduit/pkg/web/api/status"
-	"github.com/conduitio/conduit/pkg/web/api/toproto"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	"google.golang.org/grpc"
 )
