+++ pkg/http/api/processor_v1_test.go
@@ -27,10 +27,10 @@ import (
 	processorSdk "github.com/conduitio/conduit-processor-sdk"
 	"github.com/conduitio/conduit/pkg/foundation/cerrors"
 	"github.com/conduitio/conduit/pkg/foundation/log"
+	apimock "github.com/conduitio/conduit/pkg/http/api/mock"
+	"github.com/conduitio/conduit/pkg/http/api/toproto"
 	"github.com/conduitio/conduit/pkg/inspector"
 	"github.com/conduitio/conduit/pkg/processor"
-	apimock "github.com/conduitio/conduit/pkg/web/api/mock"
-	"github.com/conduitio/conduit/pkg/web/api/toproto"
 	apiv1 "github.com/conduitio/conduit/proto/api/v1"
 	"github.com/google/uuid"
 	"github.com/matryer/is"
