+++ pkg/http/openapi/copy-swagger-api.sh
