+++ pkg/http/openapi/swagger-ui/api/v1/api.swagger.json
