+++ pkg/http/openapi/swagger-ui/favicon-16x16.png
