+++ pkg/http/openapi/swagger-ui/favicon-32x32.png
