+++ pkg/http/openapi/swagger-ui/index.html
