+++ pkg/http/openapi/swagger-ui/oauth2-redirect.html
