+++ pkg/http/openapi/swagger-ui/swagger-initializer.js
