+++ pkg/http/openapi/swagger-ui/swagger-ui-es-bundle-core.js
