+++ pkg/http/openapi/swagger-ui/swagger-ui-standalone-preset.js
