+++ pkg/http/openapi/swagger-ui/swagger-ui.js
